Refactor and cleanup ComponentsEventSource
