+++ components/global/core/io.cellery.observability.k8s.client/src/main/java/io/cellery/observability/k8s/client/ComponentPodsEventSource.java
@@ -63,11 +63,6 @@ public class ComponentPodsEventSource extends Source {
 
     private static final Logger logger = Logger.getLogger(ComponentPodsEventSource.class.getName());
 
-    private static final String ATTRIBUTE_POD_NAME = "name";
-    private static final String ATTRIBUTE_DELETION_TIMESTAMP = "deletionTimestamp";
-    private static final String ATTRIBUTE_NODE_NAME = "nodeName";
-    private static final String ATTRIBUTE_STATUS = "status";
-
     private KubernetesClient k8sClient;
     private SourceEventListener sourceEventListener;
     private List<Watch> k8sWatches;
@@ -172,21 +167,21 @@ public class ComponentPodsEventSource extends Source {
         public void eventReceived(Action action, Pod pod) {
             try {
                 Map<String, Object> attributes = new HashMap<>();
-                attributes.put(Constants.ATTRIBUTE_CELL, pod.getMetadata().getLabels().get(Constants.CELL_NAME_LABEL));
-                attributes.put(Constants.ATTRIBUTE_COMPONENT, Utils.getComponentName(pod));
-                attributes.put(ATTRIBUTE_POD_NAME, pod.getMetadata().getName());
-                attributes.put(Constants.ATTRIBUTE_CREATION_TIMESTAMP, pod.getMetadata().getCreationTimestamp() == null
+                attributes.put(Constants.Attribute.CELL, pod.getMetadata().getLabels().get(Constants.CELL_NAME_LABEL));
+                attributes.put(Constants.Attribute.COMPONENT, Utils.getComponentName(pod));
+                attributes.put(Constants.Attribute.POD_NAME, pod.getMetadata().getName());
+                attributes.put(Constants.Attribute.CREATION_TIMESTAMP, pod.getMetadata().getCreationTimestamp() == null
                         ? -1
                         : new SimpleDateFormat(Constants.K8S_DATE_FORMAT, Locale.US).parse(
                                 pod.getMetadata().getCreationTimestamp()).getTime());
-                attributes.put(ATTRIBUTE_DELETION_TIMESTAMP, pod.getMetadata().getDeletionTimestamp() == null
+                attributes.put(Constants.Attribute.DELETION_TIMESTAMP, pod.getMetadata().getDeletionTimestamp() == null
                         ? -1
                         : new SimpleDateFormat(Constants.K8S_DATE_FORMAT, Locale.US).parse(
                                 pod.getMetadata().getDeletionTimestamp()).getTime());
-                attributes.put(ATTRIBUTE_NODE_NAME,
+                attributes.put(Constants.Attribute.NODE_NAME,
                         pod.getSpec().getNodeName() == null ? "" : pod.getSpec().getNodeName());
-                attributes.put(ATTRIBUTE_STATUS, pod.getStatus().getPhase());
-                attributes.put(Constants.ATTRIBUTE_ACTION, action.toString());
+                attributes.put(Constants.Attribute.STATUS, pod.getStatus().getPhase());
+                attributes.put(Constants.Attribute.ACTION, action.toString());
 
                 sourceEventListener.onEvent(attributes, new String[0]);
                 if (logger.isDebugEnabled()) {
