+++ components/global/core/io.cellery.observability.k8s.client/src/main/java/io/cellery/observability/k8s/client/crds/CellList.java
@@ -16,16 +16,13 @@
  * under the License.
  */
 
-package io.cellery.observability.k8s.client.cells;
+package io.cellery.observability.k8s.client.crds;
 
-import io.fabric8.kubernetes.api.builder.Function;
-import io.fabric8.kubernetes.client.CustomResourceDoneable;
+import io.fabric8.kubernetes.client.CustomResourceList;
 
 /**
- * This class represents Doneable resource for Cellery cells.
+ * This class represents the List resource for Cellery cells.
  */
-public class DoneableCell extends CustomResourceDoneable<Cell> {
-    public DoneableCell(Cell resource, Function<Cell, Cell> function) {
-        super(resource, function);
-    }
+public class CellList extends CustomResourceList<CellImpl> {
+    private static final long serialVersionUID = 1L;
 }
