+++ components/global/core/io.cellery.observability.k8s.client/src/main/java/io/cellery/observability/k8s/client/crds/DoneableCell.java
@@ -16,13 +16,16 @@
  * under the License.
  */
 
-package io.cellery.observability.k8s.client.cells;
+package io.cellery.observability.k8s.client.crds;
 
-import io.fabric8.kubernetes.client.CustomResourceList;
+import io.fabric8.kubernetes.api.builder.Function;
+import io.fabric8.kubernetes.client.CustomResourceDoneable;
 
 /**
- * This class represents the List resource for Cellery cells.
+ * This class represents Doneable resource for Cellery cells.
  */
-public class CellList extends CustomResourceList<Cell> {
-    private static final long serialVersionUID = 1L;
+public class DoneableCell extends CustomResourceDoneable<CellImpl> {
+    public DoneableCell(CellImpl resource, Function<CellImpl, CellImpl> function) {
+        super(resource, function);
+    }
 }
