+++ components/global/core/io.cellery.observability.k8s.client/src/main/java/io/cellery/observability/k8s/client/crds/model/HTTP.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package io.cellery.observability.k8s.client.cells.model;
+package io.cellery.observability.k8s.client.crds.model;
 
 import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
@@ -26,48 +26,86 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import io.fabric8.kubernetes.api.model.KubernetesResource;
-import io.fabric8.kubernetes.api.model.ObjectMeta;
 import java.util.HashMap;
 import java.util.Map;
 
 /**
- * This represents the serializable class for ServicesTemplate in cell yaml.
+ * This represents the serializable class for HTTP in cell yaml.
  * */
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
-        "metadata",
-        "spec"
+        "authenticate",
+        "backend",
+        "context",
+        "definitions",
+        "global"
 })
-@JsonDeserialize(using = com.fasterxml.jackson.databind.JsonDeserializer.None.class)
-public class ServicesTemplate implements KubernetesResource {
+@JsonDeserialize
+public class HTTP implements KubernetesResource {
 
     private static final long serialVersionUID = 1L;
 
-    @JsonProperty("metadata")
-    private ObjectMeta metadata;
-    @JsonProperty("spec")
-    private ServicesTemplateSpec spec;
+    @JsonProperty("authenticate")
+    private Boolean authenticate;
+    @JsonProperty("backend")
+    private String backend;
+    @JsonProperty("context")
+    private String context;
+    @JsonProperty("definitions")
+    private Object definitions;
+    @JsonProperty("global")
+    private Boolean global;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
 
-    @JsonProperty("metadata")
-    public ObjectMeta getMetadata() {
-        return metadata;
+    @JsonProperty("authenticate")
+    public Boolean getAuthenticate() {
+        return authenticate;
     }
 
-    @JsonProperty("metadata")
-    public void setMetadata(ObjectMeta metadata) {
-        this.metadata = metadata;
+    @JsonProperty("authenticate")
+    public void setAuthenticate(Boolean authenticate) {
+        this.authenticate = authenticate;
     }
 
-    @JsonProperty("spec")
-    public ServicesTemplateSpec getSpec() {
-        return spec;
+    @JsonProperty("backend")
+    public String getBackend() {
+        return backend;
     }
 
-    @JsonProperty("spec")
-    public void setSpec(ServicesTemplateSpec spec) {
-        this.spec = spec;
+    @JsonProperty("backend")
+    public void setBackend(String backend) {
+        this.backend = backend;
+    }
+
+    @JsonProperty("context")
+    public String getContext() {
+        return context;
+    }
+
+    @JsonProperty("context")
+    public void setContext(String context) {
+        this.context = context;
+    }
+
+    @JsonProperty("definitions")
+    public Object getDefinitions() {
+        return definitions;
+    }
+
+    @JsonProperty("definitions")
+    public void setDefinitions(Object definitions) {
+        this.definitions = definitions;
+    }
+
+    @JsonProperty("global")
+    public Boolean getGlobal() {
+        return global;
+    }
+
+    @JsonProperty("global")
+    public void setGlobal(Boolean global) {
+        this.global = global;
     }
 
     @JsonAnyGetter
@@ -79,5 +117,4 @@ public class ServicesTemplate implements KubernetesResource {
     public void setAdditionalProperty(String name, Object value) {
         this.additionalProperties.put(name, value);
     }
-
 }
