+++ components/global/core/io.cellery.observability.k8s.client/src/main/java/io/cellery/observability/k8s/client/crds/model/STSTemplate.java
@@ -16,28 +16,59 @@
  * under the License.
  */
 
-package io.cellery.observability.k8s.client.cells.model;
+package io.cellery.observability.k8s.client.crds.model;
 
 import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import io.fabric8.kubernetes.api.model.KubernetesResource;
+import io.fabric8.kubernetes.api.model.ObjectMeta;
 import java.util.HashMap;
 import java.util.Map;
 
 /**
- * This represents the serializable class for STSTemplate spec in cell yaml.
+ * This represents the serializable class for STSTemplate in cell yaml.
  * */
 @JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonDeserialize(using = com.fasterxml.jackson.databind.JsonDeserializer.None.class)
-public class STSTemplateSpec implements KubernetesResource {
+@JsonPropertyOrder({
+        "metadata",
+        "spec"
+})
+@JsonDeserialize
+public class STSTemplate implements KubernetesResource {
 
     private static final long serialVersionUID = 1L;
 
+    @JsonProperty("metadata")
+    private ObjectMeta metadata;
+    @JsonProperty("spec")
+    private STSTemplateSpec spec;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
+
+    @JsonProperty("metadata")
+    public ObjectMeta getMetadata() {
+        return metadata;
+    }
+
+    @JsonProperty("metadata")
+    public void setMetadata(ObjectMeta metadata) {
+        this.metadata = metadata;
+    }
+
+    @JsonProperty("spec")
+    public STSTemplateSpec getSpec() {
+        return spec;
+    }
+
+    @JsonProperty("spec")
+    public void setSpec(STSTemplateSpec spec) {
+        this.spec = spec;
+    }
 
     @JsonAnyGetter
     public Map<String, Object> getAdditionalProperties() {
@@ -48,5 +79,5 @@ public class STSTemplateSpec implements KubernetesResource {
     public void setAdditionalProperty(String name, Object value) {
         this.additionalProperties.put(name, value);
     }
-}
 
+}
