+++ components/global/core/io.cellery.observability.k8s.client/src/main/java/io/cellery/observability/k8s/client/crds/model/STSTemplateSpec.java
@@ -16,71 +16,28 @@
  * under the License.
  */
 
-package io.cellery.observability.k8s.client.cells.model;
+package io.cellery.observability.k8s.client.crds.model;
 
 import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import io.fabric8.kubernetes.api.model.KubernetesResource;
-
 import java.util.HashMap;
 import java.util.Map;
 
 /**
- * This is used to represent TCP Ingress_Type
- */
+ * This represents the serializable class for STSTemplate spec in cell yaml.
+ * */
 @JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonPropertyOrder({
-        "port",
-        "backendHost",
-        "backendPort"
-})
-@JsonDeserialize(using = com.fasterxml.jackson.databind.JsonDeserializer.None.class)
-public class TCP implements KubernetesResource {
+@JsonDeserialize
+public class STSTemplateSpec implements KubernetesResource {
+
     private static final long serialVersionUID = 1L;
 
-    @JsonProperty("port")
-    private Integer port;
-    @JsonProperty("backendHost")
-    private String backendHost;
-    @JsonProperty("backendPort")
-    private Integer backendPort;
     @JsonIgnore
     private Map<String, Object> additionalProperties = new HashMap<String, Object>();
-
-    @JsonProperty("port")
-    public Integer getPort() {
-        return port;
-    }
-
-    @JsonProperty("port")
-    public void setPort(Integer port) {
-        this.port = port;
-    }
-
-    @JsonProperty("backendHost")
-    public String getBackendHost() {
-        return backendHost;
-    }
-
-    @JsonProperty("backendHost")
-    public void setBackendHost(String backendHost) {
-        this.backendHost = backendHost;
-    }
-
-    @JsonProperty("backendPort")
-    public Integer getBackendPort() {
-        return backendPort;
-    }
-
-    @JsonProperty("backendPort")
-    public void setBackendPort(Integer backendPort) {
-        this.backendPort = backendPort;
-    }
 
     @JsonAnyGetter
     public Map<String, Object> getAdditionalProperties() {
@@ -91,5 +48,5 @@ public class TCP implements KubernetesResource {
     public void setAdditionalProperty(String name, Object value) {
         this.additionalProperties.put(name, value);
     }
-
 }
+
